Split `nano/messages` into separate files per message type
