+++ nano/messages/bulk_push.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <nano/messages/message.hpp>
+
+namespace nano
+{
+class bulk_push final : public message
+{
+public:
+	explicit bulk_push (nano::network_constants const & constants);
+	explicit bulk_push (nano::message_header const &);
+	void serialize (nano::stream &) const override;
+	bool deserialize (nano::stream &);
+	void visit (nano::message_visitor &) const override;
+
+public: // Logging
+	void operator() (nano::object_stream &) const override;
+};
+}
