+++ nano/messages/common.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+#include <nano/lib/assert.hpp>
+#include <nano/lib/stream.hpp>
+
+namespace nano
+{
+struct empty_payload
+{
+	void serialize (nano::stream &) const
+	{
+		debug_assert (false);
+	}
+	void deserialize (nano::stream &)
+	{
+		debug_assert (false);
+	}
+	void operator() (nano::object_stream &) const
+	{
+		debug_assert (false);
+	}
+};
+}
