+++ nano/messages/message_visitor.hpp
@@ -1,0 +1,75 @@
+#pragma once
+
+#include <nano/messages/asc_pull.hpp>
+#include <nano/messages/bulk_pull.hpp>
+#include <nano/messages/bulk_pull_account.hpp>
+#include <nano/messages/bulk_push.hpp>
+#include <nano/messages/confirm.hpp>
+#include <nano/messages/frontier_req.hpp>
+#include <nano/messages/keepalive.hpp>
+#include <nano/messages/node_id_handshake.hpp>
+#include <nano/messages/publish.hpp>
+#include <nano/messages/telemetry.hpp>
+
+namespace nano
+{
+class message_visitor
+{
+public:
+	virtual ~message_visitor () = default;
+
+	virtual void keepalive (nano::keepalive const & message)
+	{
+		default_handler (message);
+	};
+	virtual void publish (nano::publish const & message)
+	{
+		default_handler (message);
+	}
+	virtual void confirm_req (nano::confirm_req const & message)
+	{
+		default_handler (message);
+	}
+	virtual void confirm_ack (nano::confirm_ack const & message)
+	{
+		default_handler (message);
+	}
+	virtual void bulk_pull (nano::bulk_pull const & message)
+	{
+		default_handler (message);
+	}
+	virtual void bulk_pull_account (nano::bulk_pull_account const & message)
+	{
+		default_handler (message);
+	}
+	virtual void bulk_push (nano::bulk_push const & message)
+	{
+		default_handler (message);
+	}
+	virtual void frontier_req (nano::frontier_req const & message)
+	{
+		default_handler (message);
+	}
+	virtual void node_id_handshake (nano::node_id_handshake const & message)
+	{
+		default_handler (message);
+	}
+	virtual void telemetry_req (nano::telemetry_req const & message)
+	{
+		default_handler (message);
+	}
+	virtual void telemetry_ack (nano::telemetry_ack const & message)
+	{
+		default_handler (message);
+	}
+	virtual void asc_pull_req (nano::asc_pull_req const & message)
+	{
+		default_handler (message);
+	}
+	virtual void asc_pull_ack (nano::asc_pull_ack const & message)
+	{
+		default_handler (message);
+	}
+	virtual void default_handler (nano::message const &){};
+};
+}
