+++ nano/messages/uniquers.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <nano/lib/blocks.hpp>
+#include <nano/lib/numbers.hpp>
+#include <nano/lib/numbers_templ.hpp>
+#include <nano/lib/uniquer.hpp>
+#include <nano/lib/vote.hpp>
+
+namespace nano
+{
+using block_uniquer = nano::uniquer<nano::uint256_union, nano::block>;
+using vote_uniquer = nano::uniquer<nano::block_hash, nano::vote>;
+}
